store/wizard: extract output types into output subslice
Follow the pattern established by compliance, details, and filesystem
subslices so output state has its own module boundary.
