+++ src/Components/CreateImageWizard/steps/Oscap/tests/Oscap.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { screen } from '@testing-library/react';
 
-import { initialState } from '@/store/slices/wizard';
+import { combinedInitialState as initialState } from '@/store/slices/wizard';
 import { clickWithWait, createUser, renderWithRedux } from '@/test/testUtils';
 
 import OscapStep from '../index';
