+++ src/store/slices/wizard/index.ts
@@ -6,3 +6,4 @@ export * from './types';
 export * from './compliance/';
 export * from './details';
 export * from './filesystem';
+export * from './output';
