+++ src/store/slices/wizard/output/types.ts
@@ -1,0 +1,17 @@
+import {
+  BootcDistributionItem,
+  Distributions,
+  ImageRequest,
+  ImageTypes,
+} from '@/store/api/backend';
+
+export type ImageSource = string;
+
+export type OutputSlice = {
+  imageSource?: ImageSource | undefined;
+  isoPayloadReference?: string | undefined;
+  bootcDistributions: BootcDistributionItem[];
+  architecture: ImageRequest['architecture'];
+  distribution: Distributions;
+  imageTypes: ImageTypes[];
+};
